Move acquisition and flow graph data structures to own lib
Move them to a custom lib, to share them with the UI, and among
different workers.
