+++ src/service/acquisition/CMakeLists.txt
@@ -1,5 +1,5 @@
-add_library(od_acquisition INTERFACE acqWorker.hpp daq_api.hpp)
-target_include_directories(od_acquisition INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/.)
-target_link_libraries(od_acquisition INTERFACE majordomo disruptor project_options project_warnings)
+add_library(od_acquisition_worker INTERFACE acqWorker.hpp)
+target_include_directories(od_acquisition_worker INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/.)
+target_link_libraries(od_acquisition_worker INTERFACE od_acquisition majordomo disruptor project_options project_warnings)
 
 add_subdirectory(test)
