+++ src/service/acquisition/test/CMakeLists.txt
@@ -1,3 +1,3 @@
 add_executable(acqWorkerTest acqWorker_test.cpp)
 target_include_directories(acqWorkerTest INTERFACE ../include)
-target_link_libraries(acqWorkerTest PRIVATE fmt ut od_acquisition client zmq)
+target_link_libraries(acqWorkerTest PRIVATE fmt ut od_acquisition_worker client zmq)
